# Serialize the payload before running onSend hooks