--- a/lib/reply.js
+++ b/lib/reply.js
@@ -41,8 +41,39 @@ Reply.prototype.send = function (payload) {
     return
   }
 
+  if (payload === undefined) {
+    onSendHook(this, payload)
+    return
+  }
+
+  const contentType = this.res.getHeader('Content-Type')
+  const contentTypeNotSet = contentType === undefined
+
+  if (payload !== null) {
+    if (Buffer.isBuffer(payload) || typeof payload.pipe === 'function') {
+      if (contentTypeNotSet) {
+        this.res.setHeader('Content-Type', 'application/octet-stream')
+      }
+      onSendHook(this, payload)
+      return
+    }
+
+    if (contentTypeNotSet && typeof payload === 'string') {
+      this.res.setHeader('Content-Type', 'text/plain')
+      onSendHook(this, payload)
+      return
+    }
+  }
+
+  if (this._serializer) {
+    payload = this._serializer(payload)
+  } else if (contentTypeNotSet || contentType === 'application/json') {
+    this.res.setHeader('Content-Type', 'application/json')
+    payload = serialize(this.context, payload, this.res.statusCode)
+    flatstr(payload)
+  }
+
   onSendHook(this, payload)
-  return
 }
 
 Reply.prototype.header = function (key, value) {
@@ -122,40 +153,19 @@ function wrapOnSendEnd (err, payload) {
 }
 
 function onSendEnd (reply, payload) {
-  if (payload === undefined) {
+  if (payload === undefined || payload === null) {
     reply.sent = true
     reply.res.end()
     return
   }
 
-  var contentType = reply.res.getHeader('Content-Type')
-  if (payload !== null && typeof payload.pipe === 'function') {
-    if (!contentType) {
-      reply.res.setHeader('Content-Type', 'application/octet-stream')
-    }
+  if (typeof payload.pipe === 'function') {
     pump(payload, reply.res, pumpCallback(reply))
     return
   }
 
-  var isPayloadString = typeof payload === 'string'
-
-  if (!contentType && isPayloadString) {
-    reply.res.setHeader('Content-Type', 'text/plain')
-  } else if (Buffer.isBuffer(payload)) {
-    if (!contentType) {
-      reply.res.setHeader('Content-Type', 'application/octet-stream')
-    }
-  } else if (reply._serializer) {
-    payload = reply._serializer(payload)
-    if (typeof payload !== 'string' && !Buffer.isBuffer(payload)) {
-      throw new TypeError(`Serializer for Content-Type '${contentType}' returned invalid payload of type '${typeof payload}'. Expected a string or Buffer.`)
-    }
-  } else if (!contentType || contentType === 'application/json') {
-    reply.res.setHeader('Content-Type', 'application/json')
-    payload = serialize(reply.context, payload, reply.res.statusCode)
-    flatstr(payload)
-  } else if (!isPayloadString && !Buffer.isBuffer(payload)) {
-    throw new TypeError(`Attempted to send payload of invalid type '${typeof payload}' without serialization. Expected a string or Buffer.`)
+  if (typeof payload !== 'string' && !Buffer.isBuffer(payload)) {
+    throw new TypeError(`Attempted to send payload of invalid type '${typeof payload}'. Expected a string or Buffer.`)
   }
 
   if (!reply.res.getHeader('Content-Length')) {
@@ -214,20 +224,18 @@ function handleError (reply, error, cb) {
     return
   }
 
-  var payload = {
+  var payload = serializeError({
     error: statusCodes[statusCode + ''],
     message: error ? error.message : '',
     statusCode: statusCode
-  }
+  })
+  flatstr(payload)
   reply.res.setHeader('Content-Type', 'application/json')
 
   if (cb) {
     cb(reply, payload)
     return
   }
-
-  payload = serializeError(payload)
-  flatstr(payload)
 
   reply.res.setHeader('Content-Length', '' + Buffer.byteLength(payload))
   reply.sent = true