Serialization now happens before the `onSend` hooks, so the hooks receive the serialized payload.

This also adds a feature where an `onSend` hook can replace the payload with `null` to send a response with an empty body without having the `Content-Type` header get set to `0`. This was not possible before because `null` would either get serialized as JSON or be considered an unserialized value and cause an error to be thrown. (It would be preferable to do this by setting the payload to `undefined` in the hook, but `undefined` currently means that the payload was not changed.)

Fixes #672 

<details>
<summary>Benchmarks</summary>

**master**
```
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 3       10.65   354
[1] Req/Sec      32459.2 6549.24 37215
[1] Bytes/Sec    4.89 MB 984 kB  5.77 MB
[1]
[1] 162k requests in 5s, 24.2 MB read
...
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 2.98    10.58   376
[1] Req/Sec      32734.4 6533.01 37695
[1] Bytes/Sec    4.89 MB 984 kB  5.77 MB
[1]
[1] 164k requests in 5s, 24.4 MB read
```

**this PR**
```
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 3       10.45   341
[1] Req/Sec      32497.6 6392.17 37535
[1] Bytes/Sec    4.84 MB 964 kB  5.77 MB
[1]
[1] 162k requests in 5s, 24.2 MB read
...
[1] Stat         Avg     Stdev   Max
[1] Latency (ms) 3.01    10.63   295
[1] Req/Sec      32372.8 6272.62 37503
[1] Bytes/Sec    4.78 MB 940 kB  5.77 MB
[1]
[1] 162k requests in 5s, 24.1 MB read
```
</details>

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message and code follows [Code of conduct](https://github.com/fastify/fastify/blob/master/CODE_OF_CONDUCT.md)